requirements.txt
# requirements.txt

blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dotenv==0.9.9
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
pytesseract==0.3.13
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
ollama